recognition/tasks/dctdp/requirements.txt
dareblopy
torchjpeg
torchsummary
